+++ src/main/java/de/theidler/create_mobile_packages/items/portable_stock_ticker/SendPackage.java
@@ -1,8 +1,10 @@
 package de.theidler.create_mobile_packages.items.portable_stock_ticker;
 
 import com.simibubi.create.AllSoundEvents;
+import com.simibubi.create.content.logistics.BigItemStack;
 import com.simibubi.create.content.logistics.packagerLink.LogisticallyLinkedBehaviour;
 import com.simibubi.create.content.logistics.packagerLink.WiFiEffectPacket;
+import com.simibubi.create.content.logistics.stockTicker.PackageOrder;
 import com.simibubi.create.content.logistics.stockTicker.PackageOrderWithCrafts;
 import com.simibubi.create.foundation.advancement.AllAdvancements;
 import com.simibubi.create.foundation.networking.SimplePacketBase;
@@ -11,6 +13,9 @@ import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.network.NetworkEvent;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class SendPackage extends SimplePacketBase {
     private final PackageOrderWithCrafts order;
@@ -25,14 +30,43 @@ public class SendPackage extends SimplePacketBase {
 
     public SendPackage(FriendlyByteBuf buffer) {
         address = buffer.readUtf();
-        order = PackageOrderWithCrafts.read(buffer);
+        // Read the order from the buffer
+        int orderSize = buffer.readVarInt();
+        List<BigItemStack> stacks = new ArrayList<>();
+        for (int i = 0; i < orderSize; i++) {
+            stacks.add(new BigItemStack(
+                buffer.readItem(),
+                buffer.readVarInt()
+            ));
+        }
+        // Read the crafting information from the buffer
+        int craftingInfoSize = buffer.readVarInt();
+        List<PackageOrderWithCrafts.CraftingEntry> craftingInformation = new ArrayList<>();
+        for (int i = 0; i < craftingInfoSize; i++) {
+            craftingInformation.add(
+                PackageOrderWithCrafts.CraftingEntry.read(buffer)
+            );
+        }
+        // Create the PackageOrderWithCrafts object
+        order = new PackageOrderWithCrafts(
+            new PackageOrder(stacks),
+                craftingInformation
+        );
         encodeRequester = buffer.readBoolean();
     }
 
     @Override
     public void write(FriendlyByteBuf buffer) {
         buffer.writeUtf(address);
-        order.write(buffer);
+        buffer.writeVarInt(order.orderedStacks().stacks().size());
+        order.orderedStacks().stacks().forEach(stack -> {
+            buffer.writeItem(stack.stack);
+            buffer.writeVarInt(stack.count);
+        });
+        buffer.writeVarInt(order.orderedCrafts().size());
+        order.orderedCrafts().forEach(stack -> {
+            stack.write(buffer);
+        });
         buffer.writeBoolean(encodeRequester);
     }